requirements.txt
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
matplotlib==3.8.0
seaborn==0.13.0
pillow==10.1.0